remove obsolete classes and modify tests
we remove DBreezeSingleThreadSession and CustomThreadPoolScheduler that are no longer needed, this lead to change in some tests, and removal of a couple of tests
